requirements.txt
transformers
Pillow
fastapi